pyproject.toml
[project]
name = "styletts2-ukrainian-openai-tts-api"
version = "0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "gradio-client>=1.8.0",
    "python-dotenv>=1.1.0",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.0",
]

[tool.uv.workspace]
members = ["styletts2-ukrainian"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "ruff>=0.11.4",
]

[tool.pytest.ini_options]
disable_test_id_escaping_and_forfeit_all_rights_to_community_support = true
